#include<algorithm> -- lower_bound vector concept
Input:
The first line of the input contains the number of integers . The next line contains  integers in sorted order. The next line contains , the number of queries. Then  lines follow each containing a single integer .
If the same number is present multiple times, you have to print the first index at which it occurs.
The input is such that you always have an answer for each query.

Output Format:
For each query you have to print "Yes"(without the quotes)if the number is present and at which index it is present separated by a space.
If the number is not present you have to print "No"(without the quotes) followed by the index of the next smallest number just greater than that number.
You have to output each query in a new line.

Sample Input:
 8
 1 1 2 2 6 9 9 15
 4
 1
 4
 9
 15

Sample Output:
 Yes 1
 No 5
 Yes 6
 Yes 8